# docs/swagger.yaml
basePath: /v1/
definitions:
...
  main.activateUserRequest:
    properties:
      token:
        type: string
    type: object
  main.envelope:
    additionalProperties: {}
    type: object
  main.errorResponse:
    properties:
      error:
        type: string
    type: object
  main.registerUserRequest:
    properties:
      email:
        type: string
      facilitator:
        type: boolean
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
host: localhost:4000
info:
  contact: {}
  description: This is the API server for the Police Training application.
  title: Police Training API
  version: 1.0.0
paths:
...
  /v1/users:
    get:
      description: Retrieve a list of users with optional filters and pagination
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account and send an activation email
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.registerUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Register a new user
      tags:
      - users
  /v1/users/{id}:
    delete:
      description: Delete a user record by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.envelope'
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Delete a user
      tags:
      - users
    patch:
      description: Perform a partial update on a user record
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Update a user
      tags:
      - users
  /v1/users/activate:
    post:
      consumes:
      - application/json
      description: Activate a user account using an activation token
      parameters:
      - description: Activation token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/main.activateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/main.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Activate a user account
      tags:
      - users
  /v1/users/me:
    get:
      description: Retrieve the user record associated with the current request context
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.errorResponse'
      summary: Get current user
      tags:
      - users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"